Refactor LLVMCompileStage
This patch separates out the compilation part and delegates it to LLVMCompiler class
which now encapsulates creation of different components of compilation
toolchain.
