+++ plugins/methodinfo/src/main/java/org/qbicc/plugin/methodinfo/MethodDataEmitter.java
@@ -286,8 +286,8 @@ public class MethodDataEmitter implements Consumer<CompilationContext> {
     }
 
     private void compileMethodDataModule(CompilationContext context, Path modulePath) {
-        LLVMCompileStage compileStage = new LLVMCompileStage(isPie);
-        compileStage.compileModule(context, modulePath);
+        LLVMCompiler compiler = new LLVMCompiler(context, isPie);
+        compiler.compileModule(context, modulePath);
     }
 
     @Override
